+++ src/main/java/br/com/unitreino/Model/ModelTipoPagamento.java
@@ -8,7 +8,7 @@ public class ModelTipoPagamento {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
     private String descricao;
@@ -19,8 +19,8 @@ public class ModelTipoPagamento {
         this.descricao = descricao;
     }
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getDescricao() { return descricao; }
     public void setDescricao(String descricao) { this.descricao = descricao; }
